Added content and styled the brew page.
Added the content and styled that content on the brew page. Also
adjusted the header on the menu page to match the header on the brew
page (add and img element directly to the html rather than loading the
img as a background through css). Makes for better accessibility. Also
refined some css classes.